server/pyproject.toml
[project]
name = "playwright-test-platform"
version = "0.1.0"
description = "自然语言驱动的 Web 测试平台，基于 Playwright"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",
    "openai>=1.50.0",
    "gitpython>=3.1.0",
    "playwright>=1.48.0",
    "pytest-playwright>=0.5.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
